+++ Client/src/app/stores/AuthenticationStore.ts
@@ -7,7 +7,9 @@
 
 import { TokenResponse } from 'expo-app-auth';
 import { action, makeAutoObservable, observable } from 'mobx';
+import { Alert } from 'react-native';
 import GoogleAuth from '../authentication/GoogleAuth';
+import Strings from '../context/Strings';
 import AmphitryonDAO from '../data/AmphitryonDAO';
 import { Host, Student, UserResponse } from '../models/ApplicationTypes';
 import Utils from '../utils/Utils';
@@ -95,6 +97,8 @@ class AuthenticationStore {
       this.setUserToken(token);
       this.setIsLoading(false);
       return true;
+    } else {
+      Alert.alert(Strings.ERROR_OCCURED, Strings.ERROR_USER_LOGIN);
     }
     this.setIsLoading(false);
     return false;
