+++ Client/src/components/Calendar/styles.ts
@@ -1,0 +1,24 @@
+/**
+ * @file    styles.ts
+ * @author  Alexis Allemann & Alexandre Mottier
+ * @date    04.03.2021
+ * @brief   Calendar styles
+ */
+
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  knob: {
+    marginTop: -15,
+  },
+  meeting: {
+    width: '90%',
+  },
+  meetings: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
+});
+
+export default styles;
