--- Client/src/screens/Student/Home/styles.ts
@@ -1,24 +1,0 @@
-/**
- * @file    styles.ts
- * @author  Alexis Allemann & Alexandre Mottier
- * @date    04.03.2021
- * @brief   Student home page styles
- */
-
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  meeting: {
-    width: '90%',
-  },
-  meetings: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
-  knob: {
-    marginTop: -15,
-  },
-});
-
-export default styles;
